changing path used for uri in atk dip upload
Refs #6390

Changing the default behaviour for constructing a uri, which is written to Archivists Toolkit db (to FileVersions table).  Previously the uri would be made by taking the uri_prefix configured in admin page and adding the dip uuid + "/objects/" + file_name.
Now the uri is just:  uri_prefix + file_name.